Реорганизация структуры проекта Нормальные фигуры
